requirements.txt
llama-index
python-dotenv
Ipython
streamlit
llama-index-llms-google-genai
google-genai
pypdf
tensorflow-macos
tensorflow
tensorflow-metal
llama-index-embeddings-google-genai
llama-index-embeddings-huggingface
sentence-transformers
## build the package
-e .